refactor: move sidebar to single file
