+++ ui/src/components/layout/nav-bar/nav-bar.tsx
@@ -1,51 +1,10 @@
-import { useEffect, useRef, useState } from 'react';
-
-import { Button, Icon } from '@/components/core';
 import { useMediaQuery } from '@/hooks';
 
 import { ConnectWalletButton } from './connect-wallet-button';
 import { Logo } from './logo';
 import { NavBarStyles as Styles } from './nav-bar.styles';
-import { Navigation } from './navigations';
-
-const Sidebar: React.FC = () => {
-  const [isOpen, setIsOpen] = useState(false);
-  const sidebarRef = useRef<HTMLDivElement>(null);
-
-  const handleToggle = (): void => setIsOpen(!isOpen);
-
-  const handleNavigationClick = (): void => setIsOpen(false);
-
-  useEffect(() => {
-    if (!isOpen) return;
-    const { current } = sidebarRef;
-    if (!current) return;
-
-    const handleClickOutside = (event: MouseEvent): void => {
-      if (current && !current.contains(event.target as Node)) setIsOpen(false);
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => document.removeEventListener('mousedown', handleClickOutside);
-  }, [isOpen, sidebarRef]);
-
-  return (
-    <>
-      <Button
-        onClick={handleToggle}
-        css={{ gridArea: 'menu', fontSize: '$lg' }}
-      >
-        <Icon name="menu" />
-      </Button>
-
-      <Styles.Sidebar.Backdrop open={isOpen} />
-
-      <Styles.Sidebar.Content open={isOpen} ref={sidebarRef}>
-        <Navigation stacked onClick={handleNavigationClick} />
-      </Styles.Sidebar.Content>
-    </>
-  );
-};
+import { Navigation } from './navigation';
+import { Sidebar } from './sidebar';
 
 export const NavBar: React.FC = () => {
   const enableSidebar = useMediaQuery('(max-width: 540px)');
@@ -55,7 +14,6 @@ export const NavBar: React.FC = () => {
       <Styles.Content>
         <Logo />
         <ConnectWalletButton />
-
         {enableSidebar ? <Sidebar /> : <Navigation />}
       </Styles.Content>
     </Styles.Container>
